+++ C03/ex03/ft_strncat.c
@@ -6,20 +6,20 @@
 /*   By: sawijnbe <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/14 17:17:09 by sawijnbe          #+#    #+#             */
-/*   Updated: 2025/07/14 19:41:13 by sawijnbe         ###   ########.fr       */
+/*   Updated: 2025/07/14 21:32:22 by sawijnbe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 char	*ft_strncat(char *dest, char *src, unsigned int nb)
 {
-	int	i;
-	int	j;
+	unsigned int	i;
+	unsigned int	j;
 
 	i = 0;
 	while (dest[i])
 		i++;
 	j = -1;
-	while (src[++j] && j < (int)nb)
+	while (src[++j] && j < nb)
 		dest[i++] = src[j];
 	dest[i] = 0;
 	return (dest);
